polytope class ok, hitandrun no

+++ hitandrun/polytope.py
@@ -6,13 +6,14 @@ Oct 2018
 """
 import numpy as np
 from numba import jitclass
-from numba import int64, float32
+from numba import int64, float64
+
 spec = [
-    ("A", float32[:, :]),
-    ("b", float32[:]),
+    ("A", float64[:, :]),
+    ("b", float64[:]),
     ("dim", int64),
     ("nplanes", int64),
-    ("auxiliar_points", float32[:, :])
+    ("auxiliar_points", float64[:, :])
 ]
 
 
@@ -40,7 +41,7 @@ class Polytope(object):
         self.b = b
         self.dim = A.shape[1]
         self.nplanes = A.shape[0]
-        self.auxiliar_points = np.zeros(shape=A.shape, dtype=np.float32)
+        self.auxiliar_points = np.zeros(shape=A.shape, dtype=np.float64)
         self._find_auxiliar_points_in_planes()
 
     def check_inside(self, point):
@@ -51,14 +52,14 @@ class Polytope(object):
 
     def _find_auxiliar_points_in_planes(self):
         """Find an auxiliar point for each plane."""
-        auxiliar_points = np.zeros((self.nplanes, self.dim), dtype=np.float32)
+        auxiliar_points = np.zeros((self.nplanes, self.dim), dtype=np.float64)
         for i in range(self.nplanes):
             auxiliar_points[i] = self._find_auxiliar_point(self.A[i], self.b[i])
         self.auxiliar_points = auxiliar_points
 
     def _find_auxiliar_point(self, Ai, bi):
         """Find an auxiliar point for one plane."""
-        p = np.zeros(self.dim, dtype=np.float32)
+        p = np.zeros(self.dim, dtype=np.float64)
         j = np.argmax(Ai != 0)
         p[j] = bi / Ai[j]
         return p
